Add clean option and improve package sorting
The update introduces a "--clean" command-line option for the users to remove unused references from the project files. It also improves the sorting of package references by using the lower-case version of the PackageId for comparison. Furthermore, the term "PackageReference" is changed to "PackageVersion" to provide a more accurate descriptor.